Refactor Chat Room Component
ChatArea 분리
