+++ packages/client/src/pages/chatroom/ChatArea/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ChatArea'
