+++ packages/client/src/pages/chatroom/ChatArea/styles.ts
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+import { color } from 'styles/global'
+
+export const Container = styled.section`
+  flex:1;
+  display:flex;
+  flex-direction:column;
+  background : ${color.ROOM_COLOR};
+  overflow-y: auto;
+`
+
+export const ChatBottom = styled.div``
